+++ lib/noticed/delivery_methods/email.rb
@@ -5,6 +5,12 @@ module Noticed
         mailer.with(format).send(method.to_sym).deliver_later
       end
 
+      def self.validate!(options)
+        unless options.key?(:mailer)
+          raise ValidationError, "email delivery method requires a 'mailer' to be specified"
+        end
+      end
+
       private
 
       def mailer